unecessary cast removed

+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: atorma <atorma@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/15 18:36:08 by atorma            #+#    #+#             */
-/*   Updated: 2024/04/17 18:29:50 by atorma           ###   ########.fr       */
+/*   Updated: 2024/04/17 19:08:06 by atorma           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 
 	if (!s || !len)
 		return (NULL);
-	if ((unsigned int)ft_strlen(s) < start)
+	if (ft_strlen(s) < start)
 		return (NULL);
 	sub_len = ft_strlen(s + start);
 	if (sub_len > len)
